backend/requirements.txt
# backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
apscheduler==3.10.1
PyGithub==1.58.0
python-dotenv==1.0.0
